fix: reuse auth user for profile
